+++ app/Orm/InvitationsMapper.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\Orm;
+
+use Nextras\Orm\Mapper\Dbal\DbalMapper;
+
+/**
+ * @author Jakub Konečný
+ * @extends DbalMapper<Invitation>
+ */
+final class InvitationsMapper extends DbalMapper {
+}
+?>